# Conflicts:
#	pom.xml
#	xyz-connectors/pom.xml
#	xyz-hub-service/pom.xml
#	xyz-hub-test/pom.xml
#	xyz-jobs/pom.xml
#	xyz-jobs/xyz-job-service/pom.xml
#	xyz-jobs/xyz-job-service/src/main/java/com/here/xyz/jobs/steps/compiler/maintenance/RunSpaceMaintenance.java
#	xyz-jobs/xyz-job-steps/pom.xml
#	xyz-models/pom.xml
#	xyz-psql-connector/pom.xml
#	xyz-psql-connector/src/main/java/com/here/xyz/psql/NLConnector.java
#	xyz-psql-connector/src/main/java/com/here/xyz/psql/PSQLXyzConnector.java
#	xyz-psql-connector/src/main/java/com/here/xyz/psql/query/WriteFeatures.java
#	xyz-psql-connector/src/main/java/com/here/xyz/psql/query/branching/BranchManager.java
#	xyz-util/pom.xml

+++ xyz-hub-service/src/main/java/com/here/xyz/hub/config/dynamo/DynamoBranchConfigClient.java
@@ -29,6 +29,7 @@ import com.here.xyz.XyzSerializable;
 import com.here.xyz.XyzSerializable.Static;
 import com.here.xyz.hub.config.BranchConfigClient;
 import com.here.xyz.models.hub.Branch;
+import com.here.xyz.models.hub.Branch.DeletedBranch;
 import com.here.xyz.util.service.aws.dynamo.DynamoClient;
 import com.here.xyz.util.service.aws.dynamo.IndexDefinition;
 import io.vertx.core.Future;
@@ -88,11 +89,17 @@ public class DynamoBranchConfigClient extends BranchConfigClient {
   }
 
   @Override
-  protected Future<Void> deleteBranch(String spaceId, String branchId) {
-    return dynamoClient.executeQueryAsync(() -> {
-      branchTable.deleteItem(new DeleteItemSpec().withPrimaryKey("spaceId", spaceId, "id", branchId));
-      return null;
-    });
+  protected Future<Void> deleteBranch(String spaceId, String branchId, boolean erase) {
+    if (erase)
+      return dynamoClient.executeQueryAsync(() -> {
+        branchTable.deleteItem(new DeleteItemSpec().withPrimaryKey("spaceId", spaceId, "id", branchId));
+        return null;
+      });
+
+    //First copy the branch to the entities table before deleting it
+    return loadBranch(spaceId, branchId)
+        .compose(branch -> entityConfigClient.store(new DeletedBranch(branch)))
+        .compose(uuid -> deleteBranch(spaceId, branchId, true));
   }
 
   @Override
